- added upgrades/dbpatches/1.0beta4_to_1.0beta5/ with patches (Alexei)  - added support for disktotal[*] (Alexei)  - added support for inodetotal[*] (Alexei)  - added support for *,/,+, and - in trigger expressions (Alexei)  - removed frontends/php/chart4.php (Alexei)  - update item status to UNSUPPORTED in case if SNMP support was not included  into zabbix_suckerd (Alexei)  - added mysql_init() to DBconnect() (Alexei)  - fixed Next100 and Prev100 for case if some hosts are hidden (Alexei)  - added select for latest values (Alexei)  - delete related services if trigger is deleted (Alexei)  - fixed problem with substitution of macros for messages (Alexei)  - eliminated DBis_empty() by replacing to DBnum_rows() (Alexei)  - MAX_STRING_LEN increased to 4096 (Alexei)  - fixed zabbix_log(). Possible coredump if data contains %s, etc (Alexei)

+++ frontends/php/include/config.inc.php
@@ -433,7 +433,7 @@ where h.hostid=i.hostid and i.itemid=f.itemid and f.triggerid=$triggerid";
 // 	Replace all <float> <sign> <float> with 0
 //			echo "Expression:$expression<br>";
 			$arr="";
-			if (eregi('^((.)*)([0-9\.]+)([\&\|\>\<\=]{1})([0-9\.]+)((.)*)$', $expression, &$arr)) 
+			if (eregi('^((.)*)([0-9\.]+)([\&\|\>\<\=\+\-\*\/]{1})([0-9\.]+)((.)*)$', $expression, &$arr)) 
 			{
 //				echo "OK<br>";
 //				for($i=0;$i<50;$i++)
@@ -451,7 +451,7 @@ where h.hostid=i.hostid and i.itemid=f.itemid and f.triggerid=$triggerid";
 					$ERROR_MSG="[".$arr[5]."] is not a float";
 					return -1;
 				}
-				$expression=$arr[1]."0".$arr[6];
+				$expression=$arr[1]."(0)".$arr[6];
 	                }
 			else
 			{
@@ -460,7 +460,7 @@ where h.hostid=i.hostid and i.itemid=f.itemid and f.triggerid=$triggerid";
 
 
 // 	Replace all (float) with 0
-//			echo "Expression2:$expression<br>";
+//			echo "Expression2:[$expression]<br>";
 			$arr="";
 			if (eregi('^((.)*)(\(([0-9\.]+)\))((.)*)$', $expression, &$arr)) 
 			{
@@ -1623,7 +1623,19 @@ where h.hostid=i.hostid and i.itemid=f.itemid and f.triggerid=$triggerid";
 		}
 		$sql="delete from functions where itemid=$itemid";
 		return	DBexecute($sql);
+	}
 
+	# Delete Service definitions by triggerid
+
+	function	delete_services_by_triggerid($triggerid)
+	{
+		$sql="select serviceid from services where triggerid=$triggerid";
+		$result=DBselect($sql);
+		for($i=0;$i<DBnum_rows($result);$i++)
+		{
+			delete_service(DBget_field($result,$i,0));
+		}
+		return	TRUE;
 	}
 
 	# Delete Item definition
@@ -1741,6 +1753,11 @@ where h.hostid=i.hostid and i.itemid=f.itemid and f.triggerid=$triggerid";
 		{
 			return	$result;
 		}
+		$result=delete_services_by_triggerid($triggerid);
+		if(!$result)
+		{
+			return	$result;
+		}
 
 		$sql="delete from triggers where triggerid=$triggerid";
 		return	DBexecute($sql);
@@ -1832,7 +1849,7 @@ where h.hostid=i.hostid and i.itemid=f.itemid and f.triggerid=$triggerid";
 			$delay=DBget_field($result,$i,3);
 			$value_type=DBget_field($result,$i,4);
 
-			$itemid=add_item($description,$key,$hostid,$delay,30*24*3600,0,0,"","",$value_type,'');
+			$itemid=add_item($description,$key,$hostid,$delay,30,0,0,"","",$value_type,'');
 
 			$result2=DBselect("select triggertemplateid,description,expression from triggers_template where itemtemplateid=$itemtemplateid");
 			for($j=0;$j<DBnum_rows($result2);$j++)
